requirements.txt
numpy
pillow
tensorflow
utils
opencv
python-telegram-bot
requests
matplotlib